Review the interface, and paint numbers with the formatter that reads back
The ninth block of the closing review covers the Qt interface and the
entry point.

The dialogs painted stored values into their fields with QString::number,
which keeps six significant digits, and read the fields back on accept:
reopening the specifications of a loaded project, or touching an epsilon
in the template viewer, rounded what the file held. Every number the
interface writes into a field or a label now goes through the project's
own formatter, the one whose text reads back to the same double.

The six viewers each carried the same file dialog and format switch for
saving an image; it is one function now. The boundary union viewer kept
three display modes and two box-drawing routines that nothing called,
and is a third of its size. The template viewer held pointers into the
project's frequency and epsilon vectors and outlived them; it holds
copies. The uncertainty dialog had four different things called
rowsBuilt after an old rename, a found flag that was never set, and a
denominator loop that did not reset its validity per row.

The compiler had been reporting its warnings in Spanish, and the review
was filtering for the English word: the five it had been raising, an
initialisation order in the main window and two signed comparisons, are
fixed and the build is checked in the C locale from here on. A template
epsilon that is not a positive finite number is refused where it is
typed, the main window no longer keeps results it never read, and the
usual dead includes, empty destructors and leftover comments are gone.

+++ src/gui/frequencies_dialog.cpp
@@ -154,7 +154,7 @@ void FrequenciesDialog::on_okButton_clicked()
         }
     }
 
-    const qint32 pointCount = frequencies.size();
+    const qint32 pointCount = static_cast<qint32>(frequencies.size());
 
     m_omega = std::make_unique<Omega>(start, end, pointCount, std::move(frequencies), type);
